Add files via upload
Problem 7

By listing the first six prime numbers: 2, 3, 5, 7, 11, and 13, we can see that the 6th prime is 13.
What is the 10 001st prime number?

Note: I have quite implementation of this algorithm, which is little different from the 
one given in the overview pdf, but has really good performance. Traditional is very slow.